introduce (Button|Label|Scale)Renderer
